# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RankYatra API - Competitive Exam Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users/{userId}:
    get:
      operationId: adminGetUser
      tags: [admin]
      summary: Get user details (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"

    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update any user (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/users/{userId}/block:
    patch:
      operationId: adminBlockUser
      tags: [admin]
      summary: Block or unblock a user (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockUserBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/users/{userId}/wallet:
    patch:
      operationId: adminAdjustWallet
      tags: [admin]
      summary: Adjust user wallet balance (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustWalletBody"
      responses:
        "200":
          description: Updated wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/exams/{examId}/reward:
    post:
      operationId: adminRewardWinners
      tags: [admin]
      summary: Reward top 5 winners of a completed exam (admin only)
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RewardWinnersBody"
      responses:
        "200":
          description: Rewards distributed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RewardResult"
  /admin/exams/{examId}/solution-pdf:
    patch:
      operationId: adminUpdateSolutionPdf
      tags: [admin]
      summary: Update solution PDF URL for an exam (admin only)
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolutionPdfBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
...
components:
  schemas:
...
    Exam:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        entryFee:
          type: string
        status:
          type: string
          enum: [upcoming, live, completed]
        solutionPdfUrl:
          type: ["string", "null"]
        questionCount:
          type: integer
        registeredCount:
          type: integer
        prizePool:
          type: string
        createdAt:
          type: string
      required: [id, title, category, startTime, endTime, entryFee, status, questionCount, registeredCount, prizePool, createdAt]
...
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        walletBalance:
          type: string
        avatarUrl:
          type: ["string", "null"]
        isAdmin:
          type: boolean
        isBlocked:
          type: boolean
        createdAt:
          type: string
        totalExamsTaken:
          type: integer
        totalWinnings:
          type: string
      required: [id, name, email, walletBalance, isAdmin, isBlocked, createdAt, totalExamsTaken, totalWinnings]
    AdminUpdateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        walletBalance:
          type: string
        isAdmin:
          type: boolean
    BlockUserBody:
      type: object
      properties:
        isBlocked:
          type: boolean
      required: [isBlocked]
    AdjustWalletBody:
      type: object
      properties:
        amount:
          type: string
        type:
          type: string
          enum: [credit, debit]
        description:
          type: string
      required: [amount, type, description]
    RewardWinnersBody:
      type: object
      properties:
        prizes:
          type: array
          items:
            $ref: "#/components/schemas/PrizeEntry"
      required: [prizes]
    PrizeEntry:
      type: object
      properties:
        rank:
          type: integer
        amount:
          type: string
      required: [rank, amount]
    RewardResult:
      type: object
      properties:
        success:
          type: boolean
        rewarded:
          type: array
          items:
            $ref: "#/components/schemas/RewardedUser"
      required: [success, rewarded]
    RewardedUser:
      type: object
      properties:
        rank:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        amount:
          type: string
      required: [rank, userId, userName, amount]
    SolutionPdfBody:
      type: object
      properties:
        solutionPdfUrl:
          type: ["string", "null"]
      required: [solutionPdfUrl]
...